requirements_api.txt
# StudyMate Advanced API Requirements

# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HuggingFace & AI Models
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
huggingface-hub>=0.17.0
tokenizers>=0.15.0
sentence-transformers>=2.2.0

# Vector Database
faiss-cpu>=1.7.4

# PDF Processing
PyMuPDF>=1.23.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# HTTP & Async
httpx>=0.25.0
aiofiles>=23.2.0
aioredis>=2.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Environment & Configuration
python-dotenv>=1.0.0

# Additional Utilities
requests>=2.31.0
tqdm>=4.65.0
pillow>=10.0.0
python-magic>=0.4.27

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Optional: GPU Support
# torch[cuda]>=2.0.0
# faiss-gpu>=1.7.4
